webapp/requirements_web.txt
flask
flask-session
requests
urllib3
numpy<2.0
opencv-python-headless
Pillow
scikit-image
PyYAML